Bug 704480 - Moved /util/versions_info to a new-style middleware service. Changed dependencies accordingly.
